# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenQuality API",
    "version": "0.2.0",
    "description": "Enterprise data quality and observability platform API"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local development" }
  ],
  "paths": {
...
    "/api/v1/monitors/{id}": {
      "get": {
        "summary": "Get monitor",
        "operationId": "getMonitor",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Monitor details" }
        }
      },
      "put": {
        "summary": "Update monitor",
        "operationId": "updateMonitor",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Updated monitor" }
        }
      },
      "delete": {
        "summary": "Delete monitor",
        "operationId": "deleteMonitor",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Monitor deleted" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}